test_feature_fusion -> test_multimodal_fusion
