+++ tests/embed/test_multimodal_fusion.py
